fix: correct 32-bit Tag_val access on wasm32
Tag_val used the upstream little-endian form
((unsigned char *)val)[-sizeof(value)]. sizeof(value) is size_t
(unsigned), so -sizeof(value) is the positive constant 0xFFFFFFFC, not
-4. Native 32-bit pointer arithmetic wraps mod 2^32 and lands on -4, so
upstream gets away with it; but emcc/clang folds the positive constant
into the wasm load's static offset, which is bounds-checked on the
untruncated value and traps ("index out of bounds").

Cast the subscript to int so it is a signed -4, clang then emits wrapping i32 address arithmetic
instead of folding a negative offset into the memarg, giving the correct
address. Tag_val stays an l-value, so the Tag_set detour (and its
alloc.c/obj.c migrations) is no longer needed. Same reason Hd_val's
[-1] never trapped: its index was already signed.

+++ runtime/caml/mlvalues.h
@@ -188,7 +188,7 @@ bits  63        (64-P) (63-P)        10 9     8 7   0
 #define Tag_hp(hp) (((unsigned char *) (hp)) [sizeof(value)-1])
                                                  /* Also an l-value. */
 #else
-#define Tag_val(val) (((unsigned char *) (val)) [-sizeof(value)])
+#define Tag_val(val) (((unsigned char *) (val)) [-(int)sizeof(value)])
                                                  /* Also an l-value. */
 #define Tag_hp(hp) (((unsigned char *) (hp)) [0])
                                                  /* Also an l-value. */
